Cleaned up MainActivity
