+++ app/src/main/java/com/project/freshtomatoes/ui/MainActivity.kt
@@ -56,113 +56,12 @@ class MainActivity : ComponentActivity() {
             AppTheme(
                 useDarkTheme = true
             ) {
-
-            AuthLoginScreen()
-            }
-        }
-    }
-}
-@Composable
-fun GoogleAuthScreen(){
-    var user by remember { mutableStateOf(Firebase.auth.currentUser) }
-    val launcher = rememberFirebaseAuthLauncher(
-        onAuthComplete = { result ->
-            user = result.user
-        },
-        onAuthError = {
-            user = null
-        }
-    )
-    val token = "756539577939-lplb0tvvtbis5745q0835ndrr7bg3slc.apps.googleusercontent.com"
-    val context = LocalContext.current
-    Column {
-        if (user == null) {
-            Text("Not logged in")
-            Button(onClick = {
-                val gso =
-                    GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-                        .requestIdToken(token)
-                        .requestEmail()
-                        .build()
-                val googleSignInClient = GoogleSignIn.getClient(context, gso)
-                launcher.launch(googleSignInClient.signInIntent)
-            }) {
-                Text("Sign in via Google")
-            }
-        } else {
-            Text("Welcome ${user!!.displayName}")
-            Button(onClick = {
-                Firebase.auth.signOut()
-                user = null
-            }) {
-                Text("Sign out")
-            }
-        }
-    }
-
-
-}
-
-
-@Composable
-fun rememberFirebaseAuthLauncher(
-    onAuthComplete: (AuthResult) -> Unit,
-    onAuthError: (ApiException) -> Unit
-): ManagedActivityResultLauncher<Intent, ActivityResult> {
-    val scope = rememberCoroutineScope()
-    return rememberLauncherForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
-        val task = GoogleSignIn.getSignedInAccountFromIntent(result.data)
-        try {
-            val account = task.getResult(ApiException::class.java)!!
-            val credential = GoogleAuthProvider.getCredential(account.idToken!!, null)
-            scope.launch {
-                val authResult = Firebase.auth.signInWithCredential(credential).await()
-                onAuthComplete(authResult)
-            }
-        } catch (e: ApiException) {
-            onAuthError(e)
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun AuthLoginScreen(authViewModel: AuthViewModel =
-                        viewModel(factory= AuthViewModelFactory())
-) {
-    val userState = authViewModel.currentUser().collectAsState()
-    var email by remember {
-        mutableStateOf("")
-    }
-    Column {
-        if (userState.value == null) {
-            Text("Not logged in")
-            TextField(value = email, onValueChange = { email = it })
-            Button(onClick = {
-                authViewModel.signUp("$email", "Abcd1234!")
-            }) {
-                Text("Sign up via email")
-            }
-            Button(onClick = {
-                authViewModel.signIn("$email", "Abcd1234!")
-            }) {
-                Text("Sign in via email")
-            }
-
-        } else {
-            if (userState.value==null)
-                Text("Please sign in")
-            else
-                Text("Welcome ${userState.value!!.email}")
-            Button(onClick = {
-                authViewModel.signOut()
-            }) {
-                Text("Sign out")
-            }
-            Button(onClick = {
-                authViewModel.delete()
-            }) {
-                Text("Delete account")
+                val navController = rememberNavController()
+                CompositionLocalProvider(LocalNavController provides navController) {
+                    MainLayout {
+                        NavGraph()
+                    }
+                }
             }
         }
     }
