+++ app/src/main/java/com/project/freshtomatoes/ui/pages/Google/GoogleAuthScreen.kt
@@ -1,0 +1,57 @@
+package com.project.freshtomatoes.ui.pages.Google
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.stringResource
+import com.google.android.gms.auth.api.signin.GoogleSignIn
+import com.google.android.gms.auth.api.signin.GoogleSignInOptions
+import com.google.firebase.auth.ktx.auth
+import com.google.firebase.ktx.Firebase
+import com.project.freshtomatoes.R
+import com.project.freshtomatoes.auth.rememberFirebaseAuthLauncher
+
+@Composable
+fun GoogleAuthScreen(){
+    var user by remember { mutableStateOf(Firebase.auth.currentUser) }
+    val launcher = rememberFirebaseAuthLauncher(
+        onAuthComplete = { result ->
+            user = result.user
+        },
+        onAuthError = {
+            user = null
+        }
+    )
+    val token = stringResource(R.string.google_token)
+    val context = LocalContext.current
+    Column {
+        if (user == null) {
+            Text("Not logged in")
+            Button(onClick = {
+                val gso =
+                    GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
+                        .requestIdToken(token)
+                        .requestEmail()
+                        .build()
+                val googleSignInClient = GoogleSignIn.getClient(context, gso)
+                launcher.launch(googleSignInClient.signInIntent)
+            }) {
+                Text("Sign in via Google")
+            }
+        } else {
+            Text("Welcome ${user!!.displayName}")
+            Button(onClick = {
+                Firebase.auth.signOut()
+                user = null
+            }) {
+                Text("Sign out")
+            }
+        }
+    }
+}
