+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">FreshTomatoes</string>
+    <string name="google_token">756539577939-lplb0tvvtbis5745q0835ndrr7bg3slc.apps.googleusercontent.com</string>
 </resources>
